fix: use correct uint bits

+++ src/contracts/shibui/IShibui.sol
@@ -44,7 +44,7 @@ interface IShibui is IERC20, IERC20Permit {
 	/// @param votes Number of votes present in checkpoint.
 	struct Checkpoint {
 		uint32 fromBlock;
-		uint256 votes;
+		uint96 votes;
 	}
 
 	/*///////////////////////////////////////////////////////////////
@@ -78,12 +78,12 @@ interface IShibui is IERC20, IERC20Permit {
 	/// @notice Gets the current votes balance for `account`.
 	/// @param account The address to get votes balance.
 	/// @return The number of current votes for `account`.
-	function getCurrentVotes(address account) external view returns (uint256);
+	function getCurrentVotes(address account) external view returns (uint96);
 
 	/// @notice Determine the prior number of votes for an account as of a block number.
 	/// @dev Block number must be a finalized block or else this function will revert to prevent misinformation.
 	/// @param account The address of the account to check.
 	/// @param blockNumber The block number to get the vote balance at.
 	/// @return The number of votes the account had as of the given block.
-	function getPriorVotes(address account, uint256 blockNumber) external view returns (uint256);
+	function getPriorVotes(address account, uint256 blockNumber) external view returns (uint96);
 }
